Replaced alias in line 1626 with actual function
Changed line 1626 replacing the Alias function "mysqli_client_encoding" which has been DEPRECATED as of PHP 5.3.0 and REMOVED as of PHP 5.4.0 and replacing it with the actual function "mysqli_character_set_name". 

This will help maintain php 5 compatibility and possibly eliminate the user reported issue of "FPA never finishes" complaints when running FPA on some systems.